requirements.txt
imgurpython>=1.1.7
discord.py >=1.6.0
aiohttp>=3.7.3